spike: consolidate tests, remove proto, 2 fixes
